coredump: cleanups & pidfd extension
Hey,

The recent changes to rework coredump handling to rely on unix sockets
are in the process of being used in systemd. Yu reported on shortcoming
nameling that the signal causing the coredump was available before the
crashing process was reaped.

The previous systemd coredump container interface requires the coredump
file descriptor, and basic information including the signal number to be
sent to the container. This means we need to have the signal number
available before sending the coredump to the container.

In general, the extension makes sense and fits with the rest of the
coredump information.

In addition to this extension this fixes a bunch of the tests that were
failing and reworks the publication mechanism for exit and coredump info
retrievable via the pidfd ioctl.

To: linux-fsdevel@vger.kernel.org
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Amir Goldstein <amir73il@gmail.com>
Cc: Yu Watanabe <watanabe.yu+github@gmail.com>
Cc: Josef Bacik <josef@toxicpanda.com>
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Jann Horn <jannh@google.com>
Cc: Luca Boccassi <luca.boccassi@gmail.com>
Cc: Alexander Mikhalitsyn <alexander@mihalicyn.com>
Cc: linux-kernel@vger.kernel.org
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Jan Kara <jack@suse.cz>
Cc: Lennart Poettering <lennart@poettering.net>
Cc: Mike Yuan <me@yhndnzj.com>
Cc: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>
Signed-off-by: Christian Brauner <brauner@kernel.org>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20251026-work-coredump-signal-a72203cdf6eb",
    "prefixes": [],
    "base-branch": "master.a779e27f24aeb679969ddd1fdd7f"
  }
}